conda-requirements.txt
jinja2
pyyaml
numpy
scipy
